Require integer reception quantities

+++ backend/IQBF.Application/DTOs/Receptions/ReceptionItemRequest.cs
@@ -1,2 +1,2 @@
 namespace IQBF.Application.DTOs.Receptions;
-public sealed record ReceptionItemRequest(Guid BLId, decimal Quantity);
+public sealed record ReceptionItemRequest(Guid BLId, int Quantity);
